requirements.txt

numpy==1.21.2
pydub==0.25.1
soundfile==0.10.2
ffmpeg-python==0.2.0
